pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "funsearch-lite"
version = "0.1.0"
description = "FunSearch-style LLM-guided evolutionary search system for bin packing optimization"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "matplotlib>=3.7.0",
    "openai>=1.0.0",
    "tqdm>=4.65.0",
]

[tool.setuptools.packages.find]
include = ["funsearch_core*", "evaluator*", "llm*", "sandbox*", "store*", "experiments*", "ui*"]
exclude = ["tests*", "configs*", "artifacts*", "docs*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
